+++ src/bank/Penny.java
@@ -11,7 +11,7 @@ public class Penny extends Money
 	public  Penny(int quantity)
 	{
 		this.quantity=quantity;
-		this.total=quantity*this.value;
+		this.total=(double)quantity*this.value;
 	}
 
 	public  Penny()
